+++ Calculator/Calculator/Model/extension_BasicType.swift
@@ -1,0 +1,17 @@
+//
+//  model.swift
+//  Calculator
+//
+//  Created by Red on 2022/03/15.
+//
+
+import Foundation
+
+extension Double: CalculateItem {
+}
+
+extension String {
+    func split(with target: Character) -> [String] {
+        return self.components(separatedBy: String(target))
+    }
+}
